requirements.txt
apscheduler==3.10.4
asgiref==3.11.1
dj-database-url==3.1.0
Django==5.2.11
gunicorn==25.1.0
packaging==26.0
paho-mqtt==2.1.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.11.0
